[gg] Move Compiler.scala -> passes/Compilers.scala
