+++ sim/midas/src/main/scala/midas/Config.scala
@@ -11,6 +11,8 @@ import junctions.{NastiKey, NastiParameters}
 import freechips.rocketchip.config.{Parameters, Config, Field}
 import freechips.rocketchip.unittest.UnitTests
 
+import java.io.{File}
+
 // Provides a function to elaborate the top-level platform shim
 case object Platform extends Field[(Parameters) => PlatformShim]
 // Switches to synthesize prints and assertions
@@ -22,7 +24,7 @@ case object EnableSnapshot extends Field[Boolean]
 case object HasDMAChannel extends Field[Boolean]
 case object KeepSamplesInMem extends Field[Boolean]
 
-// MIDAS 2.0 Switches
+// Enables multi-cycle RAM model generation (as demonstrated in the ICCAD2019 paper)
 case object GenerateMultiCycleRamModels extends Field[Boolean](false)
 // User provided transforms to run before Golden Gate transformations
 // These are constructor functions accept a Parameters instance and produce a
@@ -30,6 +32,9 @@ case object GenerateMultiCycleRamModels extends Field[Boolean](false)
 case object TargetTransforms extends Field[Seq[(Parameters) => Seq[firrtl.Transform]]](Seq())
 // User provided transforms to run after Golden Gate transformations
 case object HostTransforms extends Field[Seq[(Parameters) => Seq[firrtl.Transform]]](Seq())
+
+// Directory into which output files are dumped. Set by -td when invoking the Stage
+case object OutputDir extends Field[File]
 
 class SimConfig extends Config((site, here, up) => {
   case TraceMaxLen      => 1024
